skr-event-listener: Add Request Size Limiting Middleware (#97)
* Adapt to recent KLM changes, use `ownedBy` annotation instead of label (#95)

* Adapt to recent klm changes, use ownedby annotation instead of label

* Fix linting

* update dependencies to fxi linter

* update dependencies to fix linter

* Update go lint ci

* Disable musttag

skr-event-listener: Add Request Size Limiting Middleware

* skr-event-listener: Add Request Size Limiting Middleware

* skr-event-listener: Fix Linter Issues

The musttag linter currently has a bug that lints external
dependencies of the project.
Furthermore, a few no-lints had to be added to rectify
trivial govet warnings.

* skr-event-listener: Fix Linter Issues

The musttag linter currently has a bug that lints external
dependencies of the project.
Furthermore, a few no-lints had to be added to rectify
trivial govet warnings.

* skr-event-listener: Fix Linter Issues

The musttag linter currently has a bug that lints external
dependencies of the project.
Furthermore, a few no-lints had to be added to rectify
trivial govet warnings.

* skr-event-listener: Refactor Middleware

* Fix listener tests race conditions (#98)

* fix listener tests race conditions

* implement review comments

* Adapt to recent KLM changes, use `ownedBy` annotation instead of label (#95)

* Adapt to recent klm changes, use ownedby annotation instead of label

* Fix linting

* update dependencies to fxi linter

* update dependencies to fix linter

* Update go lint ci

* Disable musttag

skr-event-listener: Add Request Size Limiting Middleware

* skr-event-listener: Fix Linter Issues

The musttag linter currently has a bug that lints external
dependencies of the project.
Furthermore, a few no-lints had to be added to rectify
trivial govet warnings.

* skr-event-listener: Refactor Middleware

---------

Co-authored-by: Jeremy Harisch <48282931+jeremyharisch@users.noreply.github.com>
Co-authored-by: khlifi411 <46890720+khlifi411@users.noreply.github.com>